+++ app/Http/Controllers/EntityController.php
@@ -147,7 +147,7 @@ class EntityController extends Controller {
                         ':entity_id' => $eid
                     ];
                 }
-                $sqlValue = \DB::select($text, $safes);
+                $sqlValue = DB::select($text, $safes);
                 // Check if only one result exists
                 if(count($sqlValue) === 1) {
                     // Get all column indices (keys) using the first row
@@ -257,7 +257,7 @@ class EntityController extends Controller {
                     ':entity_id' => $id
                 ];
             }
-            $sqlValue = \DB::select($text, $safes);
+            $sqlValue = DB::select($text, $safes);
             // Check if only one result exists
             if(count($sqlValue) === 1) {
                 // Get all column indices (keys) using the first row