+++ resources/js/components/search/Global.vue
@@ -2,11 +2,12 @@
     <multiselect
         v-model="state.entry"
         name="global-search"
+        class="multiselect-wide"
         :id="state.id"
         :object="true"
-        :label="'id'"
-        :track-by="'id'"
-        :valueProp="'id'"
+        :label="'idx'"
+        :track-by="'idx'"
+        :valueProp="'idx'"
         :mode="'single'"
         :options="query => search(query)"
         :hideSelected="false"
@@ -21,51 +22,59 @@
         :limit="limit"
         :placeholder="t('global.search')"
         @select="optionSelected">
-            <template v-slot:singlelabel="{ value }">
-                <div class="multiselect-single-label">
-                    {{ value.name }}
-                </div>
-            </template>
-            <template v-slot:option="{ option }">
-                <div class="col-3 text-center">
-                    <img :src="option.thumb_url" :alt="option.name" class="w-100" v-if="isImage(option)" />
-                    <span v-else-if="isFile(option, true)">
-                        <i class="fas fa-fw fa-file"></i>
-                    </span>
-                    <span v-else-if="isEntity(option)">
-                        <i class="fas fa-fw fa-monument"></i>
-                    </span>
-                    <span v-else-if="isBibliography(option)">
-                        <i class="fas fa-fw fa-book"></i>
-                    </span>
-                    <span v-else-if="isGeodata(option)">
-                        <i class="fas fa-fw fa-map-marked-alt"></i>
-                    </span>
-                </div>
-                <div class="col-9 ps-2">
-                    <span v-if="isFile(option)">
-                        {{ option.name }}
-                    </span>
-                    <span v-if="isBibliography(option)">
-                        {{ option.citekey }}
-                        {{ option.author }}
-                        {{ option.title }}
-                    </span>
-                    <span v-if="isEntity(option)">
-                        {{ option.name }}
-                    </span>
-                    <span v-if="isGeodata(option)">
-                        {{ option.id }}
-                    </span>
-                </div>
-            </template>
-            <template v-slot:nooptions="">
-                <div class="p-2" v-if="!!state.query" v-html="t('global.search_no_results_for', {term: state.query})">
-                </div>
-                <div class="p-1 text-muted" v-else>
-                    {{ t('global.search_no_term_info_global') }}
-                </div>
-            </template>
+        <template v-slot:singlelabel="{ value }">
+            <div class="multiselect-single-label">
+                {{ value.title }}
+            </div>
+        </template>
+        <template v-slot:option="{ option }">
+            <div class="col-3 text-center">
+                <img :src="option.thumb_url" :alt="option.title" class="w-100" v-if="isImage(option)" />
+                <span v-else-if="isFile(option, true)">
+                    <i class="fas fa-fw fa-file"></i>
+                </span>
+                <span v-else-if="isEntity(option) || isEntityAttribute(option)">
+                    <i class="fas fa-fw fa-monument"></i>
+                </span>
+                <span v-else-if="isBibliography(option)">
+                    <i class="fas fa-fw fa-book"></i>
+                </span>
+                <span v-else-if="isGeodata(option)">
+                    <i class="fas fa-fw fa-map-marked-alt"></i>
+                </span>
+            </div>
+            <div class="col-9 ps-2">
+                <span v-if="isFile(option)">
+                    {{ truncate(option.searchable.name, 50) }}
+                </span>
+                <span v-if="isBibliography(option)">
+                    <div>
+                        <div>
+                            <span class="fw-medium">{{ truncate(option.searchable.title, 50) }}</span>
+                        </div>
+                        <cite class="small">
+                            {{ truncate(option.searchable.author, 40) }} ({{ option.searchable.year }})
+                        </cite>
+                    </div>
+                </span>
+                <span v-if="isEntity(option)">
+                    {{ truncate(option.searchable.name, 50) }}
+                </span>
+                <span v-if="isEntityAttribute(option)">
+                    {{ truncate(option.title, 50) }}
+                </span>
+                <span v-if="isGeodata(option)">
+                    {{ option.title }}
+                </span>
+            </div>
+        </template>
+        <template v-slot:nooptions="">
+            <div class="p-2" v-if="!!state.query" v-html="t('global.search_no_results_for', {term: state.query})">
+            </div>
+            <div class="p-1 text-muted" v-else>
+                {{ t('global.search_no_term_info_global') }}
+            </div>
+        </template>
     </multiselect>
 </template>
 
@@ -83,7 +92,10 @@
     } from '@/api.js';
 
     import {
-        getTs
+        truncate,
+    } from '@/helpers/filters.js';
+    import {
+        getTs,
     } from '@/helpers/helpers.js';
     import {
         routeToBibliography,
@@ -119,41 +131,50 @@
                 if(!query) {
                     return await new Promise(r => r([]));
                 }
-                return await searchGlobal(query);
+                const results = await searchGlobal(query);
+                return results.map((r, i) => {
+                    return {
+                        ...r,
+                        idx: i,
+                    };
+                });
             };
             const isFile = (searchRes, withoutImages = false) => {
-                let ret = searchRes.group == 'files';
+                let ret = searchRes.type == 'files';
                 if(!ret) return false;
 
                 if(withoutImages) {
-                    ret = !searchRes.mime_type.startsWith('image/');
+                    ret = !searchRes.searchable.mime_type.startsWith('image/');
                 }
                 return ret;
             };
             const isImage = searchRes => {
-                return searchRes.group == 'files' && searchRes.mime_type.startsWith('image/');
+                return searchRes.type == 'files' && searchRes.searchable.mime_type.startsWith('image/');
             };
             const isEntity = searchRes => {
-                return searchRes.group == 'entities';
+                return searchRes.type == 'entities';
+            };
+            const isEntityAttribute = searchRes => {
+                return searchRes.type == 'entity_attribute';
             };
             const isBibliography = searchRes => {
-                return searchRes.group == 'bibliography';
+                return searchRes.type == 'bibliography';
             };
             const isGeodata = searchRes => {
-                return searchRes.group == 'geodata';
+                return searchRes.type == 'geodata';
             };
             const optionSelected = option => {
+                const obj = option.searchable;
                 if(isEntity(option)) {
-                    routeToEntity(option.id);
-                }
-                if(isFile(option)) {
-                    routeToFile(option.id);
-                }
-                if(isBibliography(option)) {
-                    routeToBibliography(option.id);
-                }
-                if(isGeodata(option)) {
-                    routeToGeodata(option.id);
+                    routeToEntity(obj.id);
+                } else if(isEntityAttribute(option)) {
+                    routeToEntity(obj.entity_id);
+                } else if(isFile(option)) {
+                    routeToFile(obj.id);
+                } else if(isBibliography(option)) {
+                    routeToBibliography(obj.id);
+                } else if(isGeodata(option)) {
+                    routeToGeodata(obj.id);
                 }
                 state.entry = {};
             };
@@ -169,11 +190,13 @@
             return {
                 t,
                 // HELPER
+                truncate,
                 // LOCAL
                 search,
                 isFile,
                 isImage,
                 isEntity,
+                isEntityAttribute,
                 isBibliography,
                 isGeodata,
                 optionSelected,